Add BF16 support to ggml backend
- Add TYPE_BF16 to FlatBuffer schema and regenerate headers
- Update Python serialize.py and ggml_backend.py to handle bfloat16
- Update C++ ggml_backend.cpp to map BF16 to GGML_TYPE_BF16
- Update tests to use bfloat16 instead of float32

The PTE size is halved (2.38GB -> 1.19GB) when using BF16.

+++ tests/test_qwen3_optimum.py
@@ -39,7 +39,7 @@ class TestQwen3OptimumExport(unittest.TestCase):
         eager_model = AutoModelForCausalLM.from_pretrained(
             model_id,
             device_map="cpu",
-            torch_dtype=torch.float32,
+            torch_dtype=torch.bfloat16,
             config=config,
             attn_implementation="sdpa",
             generation_config=GenerationConfig(
@@ -130,7 +130,7 @@ class TestQwen3OptimumExport(unittest.TestCase):
         eager_model = AutoModelForCausalLM.from_pretrained(
             model_id,
             device_map="cpu",
-            torch_dtype=torch.float32,
+            torch_dtype=torch.bfloat16,
             config=config,
             attn_implementation="sdpa",
             generation_config=GenerationConfig(
